+++ CHANGELOG.md
@@ -12,7 +12,7 @@
     - script: prevent script functions from being called in the wrong scope (ie when leaked using the `globals` object) and print an explicit error
     - containers: prevent errors with some color formats in colorWidget
     - theme: relative url not resolved correctly
-     
+
 - widgets
     - when a panel contains widgets and has `scroll` set to `true`, its value will be used to represent the scroll position for x-axis and y-axis between 0 and 1.